requirements.txt
openai>=1.12.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.0.0
pytest>=7.4.0